Major refactor of the env code.
We now use envy and serde to parse the environment. This is a lot less
error-prone than our previous approach. Deserialisers are implemented
for remote types, like Hash and Level, of course, but those are trivial
implementations that required no extra work.

This throws out 118 LoC we no longer need.

Signed-off-by: Elizabeth Myers <elizabeth.jennifer.myers@gmail.com>