Improve generation a bit.
The algorithm for generating the hash and the string has been rewritten
somewhat, although there should be no functional change.

We now use a one-step process to generate the shady string. Instead of
just copying the map to check for duplicates, we now just do it all at
once, and keep a HashSet of the elements we've seen. This should be
marginally faster.

In addition, we simply generate the separators and suffix at once as
well. This should be more efficient than constantly pushing strings.

Signed-off-by: Elizabeth Myers <elizabeth.jennifer.myers@gmail.com>